Move files in the right directories
