+++ src/ngf/json_mappings/bbox.rs
